requirements.txt
click==8.1.3
colorama==0.4.4
Flask==2.1.2
Flask-WTF==1.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
kafka-python==2.0.2
MarkupSafe==2.1.1
Pillow==9.1.0
psycopg2-binary==2.9.3
psycopg2==2.9.3
Werkzeug==2.1.2
WTForms==3.0.1
